Added stub of Activity to manage Place details plus additional files
Added basic stub of PlaceDetailsActivity :
the Activity is quite simple, but it is needed to add multiple images
for each place, so as to put them inside the HorizontalScrollView at the
bottom of the Activity.
So it is needed to modify slightly the JSON file, and the JsonUtils
class that loads JSON data.

Added activity_place_details.xml :
it is just the description of the layout of PlaceDetailsActivity.java.

Added PlacesDataContainer.java :
this class is a singleton that loads the map containing data of all
places, it also provides methods for accessing this data.
Usage of this class, instead of reading data in each Activity improves
speed and reduces memory usage.

Removed support library
Changed Activities from 'extends ActionBarActivity' to 'extends
Activity' :
using the support library and extending ActionBarActivity allows us to
support older devices, but it adds complexity to the Project.
I think we can just set the minimum sdk to 11, so the ActionBar is
already included in class Activity.

I also had to change a few lines in the manifest file and the main.xml
menu file, because they were not compiling correctly.